pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "intentfoundry"
version = "0.1.0"
description = "Open source human-AI collaboration framework — spec and architecture at the centre"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Tanoshii Computing", email = "hello@tanoshii-computing.com" }]
keywords = ["ai", "agile", "specification", "architecture", "human-ai-collaboration"]

dependencies = [
    # API framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    # Database
    "sqlalchemy>=2.0.30",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",          # PostgreSQL async driver
    "aiosqlite>=0.20.0",         # SQLite async driver (dev/test)
    "redis>=5.0.4",              # Redis client
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    # Validation & serialisation
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    # HTTP & async utilities
    "httpx>=0.27.0",
    "anyio>=4.3.0",
    # Utilities
    "python-ulid>=2.2.0",        # ULID for sortable unique IDs
    "structlog>=24.1.0",         # Structured logging
    "tenacity>=8.3.0",           # Retry logic
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
